Move violations collection into response.
